requirements.txt
google-genai
google-generativeai
openai
anthropic
python-dotenv
PyMuPDF
Pillow

